Move all chain agnostic code from block dispatchers to node core
