+++ packages/node-core/src/indexer/blockDispatcher/block-dispatcher.ts
@@ -1,0 +1,218 @@
+// Copyright 2020-2021 OnFinality Limited authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+import { getHeapStatistics } from 'v8';
+import {OnApplicationShutdown} from '@nestjs/common';
+import {EventEmitter2} from '@nestjs/event-emitter';
+import {profilerWrap} from '@subql/node-core/profiler';
+import {last} from 'lodash';
+import {Sequelize, Transaction} from 'sequelize';
+import {NodeConfig} from '../../configure';
+import {IndexerEvent} from '../../events';
+import {getLogger} from '../../logger';
+import {Queue, AutoQueue, delay, memoryLock, waitForBatchSize} from '../../utils';
+import {DynamicDsService} from '../dynamic-ds.service';
+import {PoiService} from '../poi.service';
+import { SmartBatchService } from '../smartBatch.service';
+import {StoreService} from '../store.service';
+import {StoreCacheService} from '../storeCache';
+import {IProjectNetworkConfig, IProjectService, ISubqueryProject} from '../types';
+import {BaseBlockDispatcher, ProcessBlockResponse} from './base-block-dispatcher';
+
+const logger = getLogger('BlockDispatcherService');
+
+type BatchBlockFetcher<B> = (heights: number[]) => Promise<B[]>;
+
+/**
+ * @description Intended to behave the same as WorkerBlockDispatcherService but doesn't use worker threads or any parallel processing
+ */
+export abstract class BlockDispatcher<B, DS>
+  extends BaseBlockDispatcher<Queue<number>>
+  implements OnApplicationShutdown
+{
+  private processQueue: AutoQueue<void>;
+
+  private fetchBlocksBatches: BatchBlockFetcher<B>;
+
+  private fetching = false;
+  private isShutdown = false;
+
+  protected abstract indexBlock(block: B): Promise<ProcessBlockResponse>;
+  protected abstract getBlockHeight(block: B): number;
+  protected abstract prepareTx(tx: Transaction): void;
+
+  constructor(
+    nodeConfig: NodeConfig,
+    eventEmitter: EventEmitter2,
+    projectService: IProjectService,
+    smartBatchService: SmartBatchService,
+    storeService: StoreService,
+    storeCacheService: StoreCacheService,
+    private sequelize: Sequelize,
+    poiService: PoiService,
+    project: ISubqueryProject<IProjectNetworkConfig>,
+    dynamicDsService: DynamicDsService<DS>,
+    fetchBlocksBatches: BatchBlockFetcher<B>
+  ) {
+    super(
+      nodeConfig,
+      eventEmitter,
+      project,
+      projectService,
+      new Queue(nodeConfig.batchSize * 3),
+      smartBatchService,
+      storeService,
+      storeCacheService,
+      poiService,
+      dynamicDsService
+    );
+    this.processQueue = new AutoQueue(nodeConfig.batchSize * 3);
+
+    if (this.nodeConfig.profiler) {
+      this.fetchBlocksBatches = profilerWrap(fetchBlocksBatches, 'SubstrateUtil', 'fetchBlocksBatches');
+    } else {
+      this.fetchBlocksBatches = fetchBlocksBatches;
+    }
+  }
+
+  // eslint-disable-next-line @typescript-eslint/require-await
+  async init(onDynamicDsCreated: (height: number) => Promise<void>): Promise<void> {
+    this.onDynamicDsCreated = onDynamicDsCreated;
+    const blockAmount = await this.projectService.getProcessedBlockCount();
+    this.setProcessedBlockCount(blockAmount ?? 0);
+  }
+
+  onApplicationShutdown(): void {
+    this.isShutdown = true;
+    this.processQueue.abort();
+  }
+
+  enqueueBlocks(cleanedBlocks: number[], latestBufferHeight?: number): void {
+    // // In the case where factors of batchSize is equal to bypassBlock or when cleanedBatchBlocks is []
+    // // to ensure block is bypassed, latestBufferHeight needs to be manually set
+    // If cleanedBlocks = []
+    if (!!latestBufferHeight && !cleanedBlocks.length) {
+      this.latestBufferedHeight = latestBufferHeight;
+      return;
+    }
+
+    logger.info(`Enqueueing blocks ${cleanedBlocks[0]}...${last(cleanedBlocks)}, total ${cleanedBlocks.length} blocks`);
+
+    this.queue.putMany(cleanedBlocks);
+
+    this.latestBufferedHeight = latestBufferHeight ?? last(cleanedBlocks);
+    void this.fetchBlocksFromQueue();
+  }
+
+  flushQueue(height: number): void {
+    super.flushQueue(height);
+    this.processQueue.flush();
+  }
+
+  private memoryleft(): number {
+    return (
+      this.smartBatchService.heapMemoryLimit() -
+      getHeapStatistics().used_heap_size
+    );
+  }
+
+  private async fetchBlocksFromQueue(): Promise<void> {
+    if (this.fetching || this.isShutdown) return;
+    // Process queue is full, no point in fetching more blocks
+    // if (this.processQueue.freeSpace < this.nodeConfig.batchSize) return;
+
+    this.fetching = true;
+
+    try {
+      while (!this.isShutdown) {
+        const blockNums = this.queue.takeMany(Math.min(this.nodeConfig.batchSize, this.processQueue.freeSpace));
+        // Used to compare before and after as a way to check if queue was flushed
+        const bufferedHeight = this._latestBufferedHeight;
+
+        // Queue is empty
+        if (!blockNums.length) {
+          // The process queue might be full so no block nums were taken, wait and try again
+          if (this.queue.size) {
+            await delay(1);
+            continue;
+          }
+          break;
+        }
+
+        if (this.memoryleft() < 0) {
+          //stop fetching until memory is freed
+          await waitForBatchSize(this.minimumHeapLimit);
+          continue;
+        }
+
+        logger.info(
+          `fetch block [${blockNums[0]},${blockNums[blockNums.length - 1]}], total ${blockNums.length} blocks`
+        );
+
+        // If specVersion not changed, a known overallSpecVer will be pass in
+        // Otherwise use api to fetch runtimes
+
+        if (memoryLock.isLocked) {
+          await memoryLock.waitForUnlock();
+        }
+
+        const blocks = await this.fetchBlocksBatches(blockNums);
+
+        this.smartBatchService.addToSizeBuffer(blocks);
+
+        // Check if the queues have been flushed between queue.takeMany and fetchBlocksBatches resolving
+        // Peeking the queue is because the latestBufferedHeight could have regrown since fetching block
+        if (bufferedHeight > this._latestBufferedHeight || this.queue.peek() < Math.min(...blockNums)) {
+          logger.info(`Queue was reset for new DS, discarding fetched blocks`);
+          continue;
+        }
+
+        const blockTasks = blocks.map((block) => async () => {
+          const height = this.getBlockHeight(block);
+          let tx: Transaction;
+          try {
+            tx = await this.sequelize.transaction();
+
+            this.preProcessBlock(height, tx);
+            this.prepareTx(tx);
+            // Inject runtimeVersion here to enhance api.at preparation
+            const processBlockResponse = await this.indexBlock(block);
+
+            await this.postProcessBlock(height, tx, processBlockResponse);
+
+            //set block to null for garbage collection
+            block = null;
+
+            await tx.commit();
+          } catch (e) {
+            await tx.rollback();
+            if (this.isShutdown) {
+              return;
+            }
+            logger.error(
+              e,
+              `failed to index block at height ${height} ${e.handler ? `${e.handler}(${e.stack ?? ''})` : ''}`
+            );
+            throw e;
+          }
+        });
+
+        // There can be enough of a delay after fetching blocks that shutdown could now be true
+        if (this.isShutdown) break;
+
+        this.processQueue.putMany(blockTasks);
+
+        this.eventEmitter.emit(IndexerEvent.BlockQueueSize, {
+          value: this.processQueue.size,
+        });
+      }
+    } catch (e) {
+      logger.error(e, 'Failed to fetch blocks from queue');
+      if (!this.isShutdown) {
+        process.exit(1);
+      }
+    } finally {
+      this.fetching = false;
+    }
+  }
+}
