+++ packages/node-core/src/indexer/index.ts
@@ -16,3 +16,4 @@ export * from './dictionary.service';
 export * from './sandbox';
 export * from './smartBatch.service';
 export * from './dynamic-ds.service';
+export * from './blockDispatcher';
