+++ packages/node/src/indexer/blockDispatcher/block-dispatcher.service.ts
@@ -1,51 +1,39 @@
 // Copyright 2020-2021 OnFinality Limited authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import { getHeapStatistics } from 'v8';
 import { Inject, Injectable, OnApplicationShutdown } from '@nestjs/common';
 import { EventEmitter2 } from '@nestjs/event-emitter';
 import {
-  getLogger,
   NodeConfig,
-  IndexerEvent,
-  delay,
-  AutoQueue,
-  Queue,
-  waitForBatchSize,
-  memoryLock,
   SmartBatchService,
   StoreCacheService,
   StoreService,
   IProjectService,
   PoiService,
-  BaseBlockDispatcher,
+  BlockDispatcher,
+  ProcessBlockResponse
 } from '@subql/node-core';
-import { last } from 'lodash';
 import { Sequelize, Transaction } from 'sequelize';
-import { SubqueryProject } from '../../configure/SubqueryProject';
-import * as SubstrateUtil from '../../utils/substrate';
+import {
+  SubqlProjectDs,
+  SubqueryProject,
+} from '../../configure/SubqueryProject';
 import { ApiService } from '../api.service';
 import { DynamicDsService } from '../dynamic-ds.service';
 import { IndexerManager } from '../indexer.manager';
 import { RuntimeService } from '../runtime/runtimeService';
+import { BlockContent } from '../types';
 import { UnfinalizedBlocksService } from '../unfinalizedBlocks.service';
-
-const logger = getLogger('BlockDispatcherService');
 
 /**
  * @description Intended to behave the same as WorkerBlockDispatcherService but doesn't use worker threads or any parallel processing
  */
 @Injectable()
 export class BlockDispatcherService
-  extends BaseBlockDispatcher<Queue<number>>
+  extends BlockDispatcher<BlockContent, SubqlProjectDs>
   implements OnApplicationShutdown
 {
-  private processQueue: AutoQueue<void>;
   private runtimeService: RuntimeService;
-
-  private fetching = false;
-  private isShutdown = false;
-  private fetchBlocksBatches = SubstrateUtil.fetchBlocksBatches;
 
   constructor(
     private apiService: ApiService,
@@ -56,198 +44,64 @@ export class BlockDispatcherService
     smartBatchService: SmartBatchService,
     storeService: StoreService,
     storeCacheService: StoreCacheService,
-    private sequelize: Sequelize,
+    sequelize: Sequelize,
     poiService: PoiService,
     @Inject('ISubqueryProject') project: SubqueryProject,
     dynamicDsService: DynamicDsService,
     private unfinalizedBlocksService: UnfinalizedBlocksService,
   ) {
+    const fetchBlockBatchesWrapped = async (
+      blockNums: number[],
+    ): Promise<BlockContent[]> => {
+      const specChanged = await this.runtimeService.specChanged(
+        blockNums[blockNums.length - 1],
+      );
+
+      // If specVersion not changed, a known overallSpecVer will be pass in
+      // Otherwise use api to fetch runtimes
+      return this.apiService.fetchBlocks(
+        blockNums,
+        specChanged ? undefined : this.runtimeService.parentSpecVersion,
+      );
+    };
+
     super(
       nodeConfig,
       eventEmitter,
-      project,
       projectService,
-      new Queue(nodeConfig.batchSize * 3),
       smartBatchService,
       storeService,
       storeCacheService,
+      sequelize,
       poiService,
+      project,
       dynamicDsService,
+      fetchBlockBatchesWrapped,
     );
-    this.processQueue = new AutoQueue(nodeConfig.batchSize * 3);
   }
 
-  // eslint-disable-next-line @typescript-eslint/require-await
   async init(
     onDynamicDsCreated: (height: number) => Promise<void>,
     runtimeService?: RuntimeService,
   ): Promise<void> {
-    this.onDynamicDsCreated = onDynamicDsCreated;
-    const blockAmount = await this.projectService.getProcessedBlockCount();
-    this.setProcessedBlockCount(blockAmount ?? 0);
+    await super.init(onDynamicDsCreated);
     this.runtimeService = runtimeService;
   }
 
-  onApplicationShutdown(): void {
-    this.isShutdown = true;
-    this.processQueue.abort();
+  protected getBlockHeight(block: BlockContent): number {
+    return block.block.block.header.number.toNumber();
   }
 
-  enqueueBlocks(cleanedBlocks: number[], latestBufferHeight?: number): void {
-    // // In the case where factors of batchSize is equal to bypassBlock or when cleanedBatchBlocks is []
-    // // to ensure block is bypassed, latestBufferHeight needs to be manually set
-    // If cleanedBlocks = []
-    if (!!latestBufferHeight && !cleanedBlocks.length) {
-      this.latestBufferedHeight = latestBufferHeight;
-      return;
-    }
+  protected prepareTx(tx: Transaction): void {
+    this.unfinalizedBlocksService.setTransaction(tx);
+  }
 
-    logger.info(
-      `Enqueueing blocks ${cleanedBlocks[0]}...${last(cleanedBlocks)}, total ${
-        cleanedBlocks.length
-      } blocks`,
+  protected async indexBlock(
+    block: BlockContent,
+  ): Promise<ProcessBlockResponse> {
+    const runtimeVersion = await this.runtimeService.getRuntimeVersion(
+      block.block,
     );
-
-    this.queue.putMany(cleanedBlocks);
-
-    this.latestBufferedHeight = latestBufferHeight ?? last(cleanedBlocks);
-    void this.fetchBlocksFromQueue().catch((e) => {
-      logger.error(e, 'Failed to fetch blocks from queue');
-      if (!this.isShutdown) {
-        process.exit(1);
-      }
-    });
-  }
-
-  flushQueue(height: number): void {
-    super.flushQueue(height);
-    this.processQueue.flush();
-  }
-
-  private memoryleft(): number {
-    return (
-      this.smartBatchService.heapMemoryLimit() -
-      getHeapStatistics().used_heap_size
-    );
-  }
-
-  private async fetchBlocksFromQueue(): Promise<void> {
-    if (this.fetching || this.isShutdown) return;
-    // Process queue is full, no point in fetching more blocks
-    // if (this.processQueue.freeSpace < this.nodeConfig.batchSize) return;
-
-    this.fetching = true;
-
-    try {
-      while (!this.isShutdown) {
-        const blockNums = this.queue.takeMany(
-          Math.min(this.processQueue.freeSpace, this.smartBatchSize),
-        );
-        // Used to compare before and after as a way to check if queue was flushed
-        const bufferedHeight = this._latestBufferedHeight;
-
-        // Queue is empty
-        if (!blockNums.length) {
-          // The process queue might be full so no block nums were taken, wait and try again
-          if (this.queue.size) {
-            await delay(1);
-            continue;
-          }
-          break;
-        }
-
-        if (this.memoryleft() < 0) {
-          //stop fetching until memory is freed
-          await waitForBatchSize(this.minimumHeapLimit);
-          continue;
-        }
-
-        logger.info(
-          `fetch block [${blockNums[0]},${
-            blockNums[blockNums.length - 1]
-          }], total ${blockNums.length} blocks`,
-        );
-
-        const specChanged = await this.runtimeService.specChanged(
-          blockNums[blockNums.length - 1],
-        );
-
-        // If specVersion not changed, a known overallSpecVer will be pass in
-        // Otherwise use api to fetch runtimes
-
-        if (memoryLock.isLocked) {
-          await memoryLock.waitForUnlock();
-        }
-
-        const blocks = await this.apiService.fetchBlocks(
-          blockNums,
-          specChanged ? undefined : this.runtimeService.parentSpecVersion,
-        );
-
-        this.smartBatchService.addToSizeBuffer(blocks);
-
-        // Check if the queues have been flushed between queue.takeMany and fetchBlocksBatches resolving
-        // Peeking the queue is because the latestBufferedHeight could have regrown since fetching block
-        if (
-          bufferedHeight > this._latestBufferedHeight ||
-          this.queue.peek() < Math.min(...blockNums)
-        ) {
-          logger.info(`${this.queue.peek()} - ${Math.min(...blockNums)}`);
-          logger.info(`Queue was reset for new DS, discarding fetched blocks`);
-          continue;
-        }
-
-        const blockTasks = blocks.map((block) => async () => {
-          const height = block.block.block.header.number.toNumber();
-          let tx: Transaction;
-          try {
-            const runtimeVersion = await this.runtimeService.getRuntimeVersion(
-              block.block,
-            );
-
-            tx = await this.sequelize.transaction();
-
-            this.preProcessBlock(height, tx);
-            this.unfinalizedBlocksService.setTransaction(tx);
-            // Inject runtimeVersion here to enhance api.at preparation
-            const processBlockResponse = await this.indexerManager.indexBlock(
-              block,
-              runtimeVersion,
-            );
-
-            await this.postProcessBlock(height, processBlockResponse);
-
-            //set block to null for garbage collection
-            block = null;
-
-            await this.storeCacheService.flushCache();
-            await tx.commit();
-          } catch (e) {
-            await tx.rollback();
-            if (this.isShutdown) {
-              return;
-            }
-            logger.error(
-              e,
-              `failed to index block at height ${height} ${
-                e.handler ? `${e.handler}(${e.stack ?? ''})` : ''
-              }`,
-            );
-            throw e;
-          }
-        });
-
-        // There can be enough of a delay after fetching blocks that shutdown could now be true
-        if (this.isShutdown) break;
-
-        this.processQueue.putMany(blockTasks);
-
-        this.eventEmitter.emit(IndexerEvent.BlockQueueSize, {
-          value: this.processQueue.size,
-        });
-      }
-    } finally {
-      this.fetching = false;
-    }
+    return this.indexerManager.indexBlock(block, runtimeVersion);
   }
 }
